Use the post_install and uninstall hooks of Sodexis developped by @atchuthan
Add automated test for post_install hook
One odoo model per file
