+++ partner_address_street3/model/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import partner
+from . import country
