+++ partner_address_street3/model/partner.py
@@ -17,16 +17,3 @@ class ResPartner(models.Model):
         fields = super(ResPartner, self)._address_fields()
         fields.append('street3')
         return fields
-
-
-class ResCountry(models.Model):
-    """Override default adresses formatting of countries"""
-    _inherit = 'res.country'
-
-    address_format = fields.Text(
-        default=(
-            "%(street)s\n%(street2)s\n%(street3)s\n"
-            "%(city)s %(state_code)s %(zip)s\n"
-            "%(country_name)s"
-        )
-    )
