+++ partner_address_street3/view/partner_view.xml
@@ -2,7 +2,7 @@
 <odoo>
 
     <record id="add_street3_in_form_view" model="ir.ui.view">
-      <field name="name">add stree3 in form view</field>
+      <field name="name">add street3 in form view</field>
       <field name="model">res.partner</field>
       <field name="inherit_id" ref="base.view_partner_form"/>
       <field name="arch" type="xml">
